luah/requirements.txt
lupa>=1.14
pygame>=2.0.0
websocket-client>=1.0.0
